requirements.txt
validators
youtube-transcript-api
langchain
langchain-community
langchain-groq
streamlit
python-dotenv
unstructured
pytube
